// frontend/src/state/ducks/events/types.ts
export enum ActionTypes {
  REQUEST_EVENT = 'REQUEST_EVENT',
  SUCCESS_EVENT = 'SUCCESS_EVENT',
  ERROR_EVENT = 'ERROR_EVENT',
  REQUEST_NEWLY_EVENT = 'REQUEST_NEWLY_EVENT',
  SUCCESS_NEWLY_EVENT = 'SUCCESS_NEWLY_EVENT',
  ERROR_NEWLY_EVENT = 'ERROR_NEWLY_EVENT',
  REQUEST_FINISHED_EVENT = 'REQUEST_FINISHED_EVENT',
  SUCCESS_FINISHED_EVENT = 'SUCCESS_FINISHED_EVENT',
  ERROR_FINISHED_EVENT = 'ERROR_FINISHED_EVENT'
}